+++ app/Bundles/Admin/Migrations/2023_10_24_000000_create_admin_roles_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('admin_roles', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('admin_user_id')->comment('用户ID');
-            $table->unsignedInteger('role_id')->comment('角色ID');
+            $table->unsignedBigInteger('admin_user_id')->comment('用户ID');
+            $table->unsignedBigInteger('role_id')->comment('角色ID');
             $table->unique(['admin_user_id', 'role_id'], 'admin_user_role_id');
             $table->comment('管理员与角色关联表');
         });
