+++ app/Bundles/Advertising/Migrations/2023_10_24_000000_create_advertising_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('advertising', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('parent_id')->default(0)->comment('类型:0广告位,其他为广告内容');
+            $table->unsignedBigInteger('parent_id')->default(0)->comment('类型:0广告位,其他为广告内容');
             $table->string('name')->default('')->comment('标题');
             $table->string('description')->default('')->comment('描述');
             $table->unsignedInteger('width')->default(0)->comment('广告宽度');
@@ -24,7 +24,7 @@ return new class extends Migration
             $table->dateTime('end_time')->comment('结束时间');
             $table->unsignedInteger('click_count')->default(0)->comment('点击量');
             $table->unsignedInteger('sort')->default(0)->comment('排序');
-            $table->unsignedInteger('status')->default(1)->comment('状态:1正常,2禁用');
+            $table->unsignedTinyInteger('status')->default(1)->comment('状态:1正常,2禁用');
             $table->timestamps();
             $table->softDeletes();
             $table->comment('广告表');
