+++ app/Bundles/Region/Migrations/2023_10_24_000000_create_regions_table.php
@@ -13,12 +13,12 @@ return new class extends Migration
     {
         Schema::create('regions', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('parent_id')->default(0)->comment('父级ID');
+            $table->unsignedBigInteger('parent_id')->default(0)->comment('父级ID');
             $table->string('name')->default('')->comment('名称');
             $table->string('pinyin')->default('')->comment('拼音');
             $table->char('first_letter')->default('')->comment('首字母');
             $table->unsignedInteger('sort')->default(0)->comment('排序');
-            $table->unsignedInteger('status')->default(1)->comment('状态:1正常,2禁用');
+            $table->unsignedTinyInteger('status')->default(1)->comment('状态:1正常,2禁用');
             $table->timestamps();
             $table->softDeletes();
             $table->comment('行政地区表');
